MVP/requirements.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.0.0
huggingface-hub==0.19.4
easyocr==1.7.1
opencv-python==4.8.1.78 